doc-knowledge-forge/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pymupdf==1.23.8
python-docx==1.1.0
markdownify==0.11.6
python-multipart==0.0.6
sqlalchemy==2.0.23
aiofiles==23.2.1
pillow==10.1.0
pydantic==2.5.2
pydantic-settings==2.1.0
